requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
python-firebase==1.2
gunicorn==20.1.0
Pillow===6.2.2
flask_cors
imageio
opencv-contrib-python==4.6.*
firebase_admin
dill